requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stripe==6.1.0
typing_extensions==4.7.1
urllib3==2.0.4

# NPM Modules
# "axios": "^1.5.0",
# "react": "^18.2.0",
# "react-dom": "^18.2.0",
# "react-scripts": "5.0.1",
# "web-vitals": "^2.1.4"
# "react-router-dom": "^6.15.0"
